cs408_project/requirements.txt
django==1.11
pillow
PyPDF2
pydub
autobahn>=0.10.9
pyOpenSSL>=0.13.1
requests>=2.8.1
Twisted>=13.2.0
txaio>=2.0.4
service_identity>=16.0.0
python-rake
cython
jumpy
nltk
jellyfish
pdfminer.six
beautifulsoup4
